requirements.txt
requests==2.32.3
python-telegram-bot==13.0
python-dotenv==0.21.0